Rename componentFactories directory
