+++ packages/framework/aqueduct/src/componentFactories/sharedComponentFactory.ts
