+++ packages/framework/aqueduct/src/index.ts
@@ -13,7 +13,7 @@
  * @packageDocumentation
  */
 
+export * from "./componentFactories";
 export * from "./components";
 export * from "./containerRuntimes";
 export * from "./containerServices";
-export * from "./helpers";
